implemented for uint(32 and 64)

+++ ch6/ex6_5/intset_test.go
@@ -250,7 +250,7 @@ func TestElem(t *testing.T) {
 	if len(x.Elem()) != 3 {
 		t.Errorf("len(x.Elem()) != 3. len(x.Elem()) is %d", len(x.Elem()))
 	}
-	y := []uint64{1, 9, 144}
+	y := []uint{1, 9, 144}
 	for i, _ := range x.Elem() {
 		if x.Elem()[i] != y[i] {
 			t.Errorf("x.Elem != {1, 9, 144}. x.Eleme[%d] is %d", i, x.Elem()[i])
